feat: Add content tags tree state + editing
This commit adds the editing functionality for content tags (minus
commuincation with the backend). You are able to select/deselect tags
from the dropdowns in each taxonomy and the tags tree structure is
updated and the state persists.

When a tag is select, all its ancestors automatically become 'implicit'
and are unselected and connot be selected. If an explicit tag is
unselected and the parent (or ancestor) no longer has explicit children,
all ancestor tags will also be removed.