fix: request only the active courses for home (#87)
By default the list of courses on the home page are
limited to "active" courses only. However, the
API query wasn't filtering for active courses, leading
to a bug showing archived courses in the list. This
commit adds the active_only filter for the first load.

Internal-ref: https://tasks.opencraft.com/browse/BB-9718