fix: Replace `Truncate.Deprecated` with CSS-based truncation (#2374)
The Truncate element as it exists has a bug where it can end up in an infinite loop when truncating to a very small size on mobiles. This makes the course outline view unresponsive on mobile and can lead to a crash.

This replaces the Truncate element with some CSS.

(cherry picked from commit 53e925e07a23ed25de6ff7a68c5160bf7d42f8c1)